+++ lib/ember_cli/runner.rb
@@ -2,17 +2,17 @@ require "open3"
 
 module EmberCli
   class Runner
-    def initialize(env: {}, out:, err:, options: {})
+    def initialize(out:, err:, env: {}, options: {})
       @env = env
-      @out = out
-      @err = err
+      @out_streams = Array(out)
+      @err_streams = Array(err)
       @options = options
     end
 
     def run(command)
-      output, status = Open3.capture2e(@env, command, @options)
+      output, status = Open3.capture2e(env, command, options)
 
-      @out.write(output)
+      write_to_out(output)
 
       [output, status]
     end
@@ -21,7 +21,7 @@ module EmberCli
       output, status = run(command)
 
       unless status.success?
-        @err.write <<-MSG.strip_heredoc
+        write_to_err <<-MSG.strip_heredoc
           ERROR: Failed command: `#{command}`
           OUTPUT:
             #{output}
@@ -31,6 +31,26 @@ module EmberCli
       end
 
       true
+    end
+
+    protected
+
+    attr_reader :env, :err_streams, :options, :out_streams
+
+    private
+
+    def write_to_out(output)
+      write(out_streams, output)
+    end
+
+    def write_to_err(output)
+      write(out_streams + err_streams, output)
+    end
+
+    def write(streams, output)
+      streams.each do |stream|
+        stream.write(output)
+      end
     end
   end
 end