+++ lib/ember_cli/runner.rb
@@ -4,18 +4,22 @@ module EmberCli
   class Runner
     def initialize(out:, err:, env: {}, options: {})
       @env = env
-      @out_streams = Array(out)
-      @err_streams = Array(err)
+      @output_streams = Array(out)
+      @error_streams = Array(err)
       @options = options
+      @threads = []
     end
 
     def run(command)
-      output, error, status = Open3.capture3(env, command, options)
+      Open3.popen3(env, command, options) do |stdin, stdout, stderr, process|
+        stdin.close
 
-      write(output, streams: out_streams)
-      write(error, streams: err_streams)
+        threads << redirect_stream_in_thread(stdout, write_to: output_streams)
+        threads << redirect_stream_in_thread(stderr, write_to: error_streams)
 
-      status
+        threads.each(&:join)
+        process.value
+      end
     end
 
     def run!(command)
@@ -28,13 +32,17 @@ module EmberCli
 
     protected
 
-    attr_reader :env, :err_streams, :options, :out_streams
+    attr_reader :env, :error_streams, :options, :output_streams, :threads
 
     private
 
-    def write(output, streams:)
-      streams.each do |stream|
-        stream.write(output)
+    def redirect_stream_in_thread(stream, write_to:)
+      Thread.new do
+        Thread.current.abort_on_exception = true
+
+        while line = stream.gets
+          write_to.each { |redirection_stream| redirection_stream.puts(line) }
+        end
       end
     end
   end