+++ lib/ember_cli/shell.rb
@@ -85,7 +85,7 @@ module EmberCli
     def runner
       Runner.new(
         options: { chdir: paths.root.to_s },
-        out: [$stdout, paths.log],
+        out: [$stdout, paths.log.open("a")],
         err: [$stderr],
         env: env,
       )