+++ spec/lib/ember_cli/runner_spec.rb
@@ -4,37 +4,33 @@ describe EmberCli::Runner do
   describe "#run!" do
     context "when the command fails" do
       it "writes output to `out` streams" do
-        output_streams = Array.new(2) { StringIO.new }
+        stdout = StringIO.new
+        logfile = StringIO.new
         runner = EmberCli::Runner.new(
           err: [],
-          out: output_streams,
+          out: [stdout, logfile],
         )
 
-        expect { runner.run!("echo 'out'; echo 'err' > /dev/stderr; exit 1") }.
+        expect { runner.run!(command_with_error(out: "out")) }.
           to raise_error(SystemExit)
 
-        ouput_strings = output_streams.each(&:rewind).map(&:read)
-
-        ouput_strings.each do |output|
-          expect(output).to eq("out\n")
-        end
+        expect(split_output_from_stream(stdout)).to eq(%w[out out])
+        expect(split_output_from_stream(logfile)).to eq(%w[out out])
       end
 
       it "writes errors to `err` streams" do
-        error_streams = Array.new(2) { StringIO.new }
+        stderr = StringIO.new
+        logfile = StringIO.new
         runner = EmberCli::Runner.new(
-          err: error_streams,
+          err: [stderr, logfile],
           out: [],
         )
 
-        expect { runner.run!("echo 'out'; echo 'err' > /dev/stderr; exit 1") }.
+        expect { runner.run!(command_with_error(err: "err")) }.
           to raise_error(SystemExit)
 
-        error_strings = error_streams.each(&:rewind).map(&:read)
-
-        error_strings.each do |error|
-          expect(error).to eq("err\n")
-        end
+        expect(split_output_from_stream(stderr)).to eq(%w[err err])
+        expect(split_output_from_stream(logfile)).to eq(%w[err err])
       end
     end
 
@@ -43,13 +39,33 @@ describe EmberCli::Runner do
       err = StringIO.new
       runner = EmberCli::Runner.new(err: [err], out: [out])
 
-      status = runner.run!("echo 'out'")
+      status = runner.run!(command("out"))
 
       [err, out].each(&:rewind)
 
       expect(status).to be_success
-      expect(err.read).to be_empty
-      expect(out.read).to eq("out\n")
+      expect(split_output_from_stream(err)).to be_empty
+      expect(split_output_from_stream(out)).to eq(%w[out])
+    end
+
+    def split_output_from_stream(stream)
+      stream.rewind
+
+      stream.read.split
+    end
+
+    def command(output)
+      "echo '#{output}'"
+    end
+
+    def command_with_error(out: "", err: "")
+      [
+        "echo '#{out}'",
+        "echo '#{err}' > /dev/stderr",
+        "echo '#{out}'",
+        "echo '#{err}' > /dev/stderr",
+        "exit 1",
+      ].join("; ")
     end
   end
 end