+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 master
 ------
 
+* Stream output instead of waiting until subprocesses finish. [#423]
 * Update `ember-cli-rails-assets` dependency. [#422]
 * Only write errors to `STDERR`. [#421]
-* Remove dependency on `tee`. Fixes bug [#417][#417]. [#420]
+* Remove dependency on `tee`. Fixes bug [#417]. [#420]
 
+[#423]: https://github.com/thoughtbot/ember-cli-rails/pull/423
 [#422]: https://github.com/thoughtbot/ember-cli-rails/pull/422
 [#421]: https://github.com/thoughtbot/ember-cli-rails/issues/421
 [#420]: https://github.com/thoughtbot/ember-cli-rails/issues/420